Sets version to 0.5.0 and adds some out-commented code in an attempt to get Stride part working
To debug Stride: Set startup project to VL.CEF.Stride, then open the VL.CEF/help/Stride overview patch manually.

Note: One needs to open the Stride overview patch manually, otherwise we get a null pointer in the beginning because the asset database is not found. Not sure what's going on there, but seems to be vvvv bug.